move filter chores to slice and fix issue of date
